+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.cs
@@ -54,7 +54,7 @@ namespace Microsoft.UI.Xaml.Controls;
 /// but the ContentSource property is not available, because there are ControlTemplates for now.
 /// </remarks>
 [ContentProperty(Name = "Content")]
-public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePoolAware
+public partial class ContentPresenter : FrameworkElement
 #if !__CROSSRUNTIME__ && !IS_UNIT_TESTS
 	, ICustomClippingElement
 #endif
@@ -63,14 +63,12 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 	private readonly BorderLayerRenderer _borderRenderer;
 #endif
 
-	private bool _firstLoadResetDone;
 	private View _contentTemplateRoot;
-	private bool _appliedTemplate;
 
 	/// <summary>
 	/// Will be set to either the result of ContentTemplateSelector or to ContentTemplate, depending on which is used
 	/// </summary>
-	private DataTemplate _dataTemplateUsedLastUpdate;
+	private FrameworkTemplate _dataTemplateUsedLastUpdate;
 
 	public ContentPresenter()
 	{
@@ -122,7 +120,7 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 			typeof(ContentPresenter),
 			new FrameworkPropertyMetadata(
 				defaultValue: null,
-				options: FrameworkPropertyMetadataOptions.None,
+				options: ContentPropertyOptions,
 				propertyChangedCallback: (s, e) => ((ContentPresenter)s)?.OnContentChanged(e.OldValue, e.NewValue)
 			)
 		);
@@ -639,90 +637,6 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 
 	#endregion
 
-	protected override void OnApplyTemplate()
-	{
-		base.OnApplyTemplate();
-
-		// Applying the template will not delete existing visuals. This will be done conditionally
-		// when the template is invalidated.
-		// Uno specific: since we don't call this early enough, we have to comment out the condition
-		// if (GetChildren().Count == 0)
-		{
-			ContentControl pTemplatedParent = TemplatedParent as ContentControl;
-
-			// Only ContentControl has the two properties below.  Other parents would just fail to bind since they don't have these
-			// two content related properties.
-			if (pTemplatedParent != null
-#if ANDROID || __IOS__
-				&& this is not NativeCommandBarPresenter // Uno specific: NativeCommandBarPresenter breaks if you inherit from the TP
-#endif
-				)
-			{
-				// bool needsRefresh = false;
-				DependencyProperty pdpTarget;
-
-				// By default Content and ContentTemplate are template are bound.
-				// If no template binding exists already then hook them up now
-				// pdpTarget = GetPropertyByIndexInline(KnownPropertyIndex::ContentPresenter_SelectedContentTemplate);
-				// IFCEXPECT(pdpTarget);
-				// if (IsPropertyDefault(pdpTarget) && !IsPropertyTemplateBound(pdpTarget))
-				// {
-				// 	const CDependencyProperty* pdpSource = pTemplatedParent->GetPropertyByIndexInline(KnownPropertyIndex::ContentControl_SelectedContentTemplate);
-				// 	IFCEXPECT(pdpSource);
-				//
-				// 	IFC(SetTemplateBinding(pdpTarget, pdpSource));
-				// 	needsRefresh = true;
-				// }
-
-				// UNO Specific: SelectedContentTemplate is not implemented, we hook ContentTemplateSelector instead
-				pdpTarget = ContentPresenter.ContentTemplateSelectorProperty;
-				global::System.Diagnostics.Debug.Assert(pdpTarget is { });
-				var store = ((IDependencyObjectStoreProvider)this).Store;
-				if (store.GetCurrentHighestValuePrecedence(pdpTarget) == DependencyPropertyValuePrecedences.DefaultValue &&
-					!store.IsPropertyTemplateBound(pdpTarget))
-				{
-					DependencyProperty pdpSource = ContentControl.ContentTemplateSelectorProperty;
-					global::System.Diagnostics.Debug.Assert(pdpSource is { });
-
-					store.SetTemplateBinding(pdpTarget, pdpSource);
-					// needsRefresh = true;
-				}
-
-				pdpTarget = ContentPresenter.ContentTemplateProperty;
-				global::System.Diagnostics.Debug.Assert(pdpTarget is { });
-				if (store.GetCurrentHighestValuePrecedence(pdpTarget) == DependencyPropertyValuePrecedences.DefaultValue &&
-					!store.IsPropertyTemplateBound(pdpTarget))
-				{
-					DependencyProperty pdpSource = ContentControl.ContentTemplateProperty;
-					global::System.Diagnostics.Debug.Assert(pdpSource is { });
-
-					store.SetTemplateBinding(pdpTarget, pdpSource);
-					// needsRefresh = true;
-				}
-
-				pdpTarget = ContentPresenter.ContentProperty;
-				global::System.Diagnostics.Debug.Assert(pdpTarget is { });
-				if (store.GetCurrentHighestValuePrecedence(pdpTarget) == DependencyPropertyValuePrecedences.DefaultValue &&
-					!store.IsPropertyTemplateBound(pdpTarget))
-				{
-					DependencyProperty pdpSource = ContentControl.ContentProperty;
-					global::System.Diagnostics.Debug.Assert(pdpSource is { });
-
-					store.SetTemplateBinding(pdpTarget, pdpSource);
-					// needsRefresh = true;
-				}
-
-				// Uno specific: uno bindings don't work this way
-				// Setting up the binding doesn't get you the values.  We need to call refresh to get the latest value
-				// for m_pContentTemplate, SelectedContentTemplate and/or m_pContent for the tests below.
-				// if (needsRefresh)
-				// {
-				// 	IFC(pTemplatedParent->RefreshTemplateBindings(TemplateBindingsRefreshType::All));
-				// }
-			}
-		}
-	}
-
 	protected virtual void OnForegroundColorChanged(Brush oldValue, Brush newValue)
 	{
 		OnForegroundColorChangedPartial(oldValue, newValue);
@@ -767,42 +681,50 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 
 	protected virtual void OnContentChanged(object oldValue, object newValue)
 	{
+#if UNO_HAS_ENHANCED_LIFECYCLE
+		//if (TemplatedParent is ContentControl contentControl)
+		//{
+		//	contentControl.ConsiderContentPresenterForContentTemplateRoot(this, newValue);
+		//}
+#endif
+
 		if (oldValue is View || newValue is View)
 		{
 			// Make sure not to reuse the previous Content as a ContentTemplateRoot (i.e., in case there's no data template)
 			// If setting Content to a new View, recreate the template
 			ContentTemplateRoot = null;
+#if UNO_HAS_ENHANCED_LIFECYCLE
+			Invalidate(true);
+#endif
+		}
+		else
+		{
+			Invalidate(false);
 		}
 
+#if !UNO_HAS_ENHANCED_LIFECYCLE
 		TrySetDataContextFromContent(newValue);
+#endif
 
 		if (IsInLiveTree)
 		{
 			TryRegisterNativeElement(oldValue, newValue);
 		}
 
+#if !UNO_HAS_ENHANCED_LIFECYCLE
 		SetUpdateTemplate();
+#endif
 	}
 
 	private void TrySetDataContextFromContent(object value)
 	{
-		if (value == null)
+		if (value == null || value is View)
 		{
 			this.ClearValue(DataContextProperty, DependencyPropertyValuePrecedences.Local);
 		}
 		else
 		{
-			if (!(value is View))
-			{
-				// If the content is not a view, we apply the content as the
-				// DataContext of the materialized content.
-				DataContext = value;
-			}
-			else
-			{
-				// Restore DataContext propagation if the content is a view
-				this.ClearValue(DataContextProperty, DependencyPropertyValuePrecedences.Local);
-			}
+			DataContext = value;
 		}
 	}
 
@@ -815,12 +737,16 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 
 	protected virtual void OnContentTemplateChanged(DataTemplate oldContentTemplate, DataTemplate newContentTemplate)
 	{
+#if UNO_HAS_ENHANCED_LIFECYCLE
+		Invalidate(true);
+#else
 		if (ContentTemplateRoot != null)
 		{
 			ContentTemplateRoot = null;
 		}
 
 		SetUpdateTemplate();
+#endif
 	}
 
 	protected virtual void OnContentTemplateSelectorChanged(DataTemplateSelector oldContentTemplateSelector, DataTemplateSelector newContentTemplateSelector)
@@ -943,52 +869,11 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 		}
 	}
 
-#if UNO_HAS_ENHANCED_LIFECYCLE
-	internal override void Enter(EnterParams @params, int depth)
-	{
-		base.Enter(@params, depth);
-
-		if (ResetDataContextOnFirstLoad() || ContentTemplateRoot == null)
-		{
-			SetUpdateTemplate();
-		}
-
-		// When the control is loaded, set the TemplatedParent
-		// as it may have been reset during the last unload.
-		SynchronizeContentTemplatedParent();
-
-#if !UNO_HAS_BORDER_VISUAL
-		UpdateBorder();
-#endif
-
-		// We do this in Enter not Loaded since Loaded is a lot more tricky
-		// (e.g. you can have Unloaded without Loaded, you can have multiple loaded events without unloaded in between, etc.)
-		if (!IsNativeHost)
-		{
-			TryRegisterNativeElement(null, Content);
-		}
-		if (IsNativeHost) // IsNativeHost can become true after the above TryRegisterNativeElement call, so this is not an if-else situation
-		{
-			AttachNativeElement();
-		}
-	}
-
-	internal override void Leave(LeaveParams @params)
-	{
-		base.Leave(@params);
-
-		if (IsNativeHost)
-		{
-			DetachNativeElement(Content);
-		}
-	}
-#endif
-
 	private protected override void OnLoaded()
 	{
 		base.OnLoaded();
 
-
+#if !UNO_HAS_ENHANCED_LIFECYCLE
 		// WinUI has some special handling for ContentPresenter and ContentControl where even though they aren't Controls,
 		// they use OnApplyTemplate. As a workaround for now, we just call OnApplyTemplate here.
 		if (!_appliedTemplate)
@@ -998,25 +883,22 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 		}
 
 
-#if !UNO_HAS_ENHANCED_LIFECYCLE
 		if (ResetDataContextOnFirstLoad() || ContentTemplateRoot == null)
 		{
 			SetUpdateTemplate();
 		}
-#endif
 
 		// When the control is loaded, set the TemplatedParent
 		// as it may have been reset during the last unload.
 		SynchronizeContentTemplatedParent();
 
-#if !UNO_HAS_ENHANCED_LIFECYCLE
 		UpdateBorder();
 
 		if (!IsNativeHost)
 		{
 			TryRegisterNativeElement(null, Content);
 		}
-		if (IsNativeHost)
+		if (IsNativeHost) // IsNativeHost can become true after the above TryRegisterNativeElement call, so this is not an if-else situation
 		{
 			AttachNativeElement();
 		}
@@ -1035,91 +917,16 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 #endif
 	}
 
-	private bool ResetDataContextOnFirstLoad()
-	{
-		if (!_firstLoadResetDone)
-		{
-			_firstLoadResetDone = true;
-
-			// This test avoids the ContentPresenter from resetting
-			// the DataContext to null (or the inherited value) and then back to
-			// the content and have two-way bindings propagating the null value
-			// back to the source.
-			if (!ReferenceEquals(DataContext, Content))
-			{
-				// On first load UWP clears the local value of a ContentPresenter.
-				// The reason for this behavior is unknown.
-				this.ClearValue(DataContextProperty, DependencyPropertyValuePrecedences.Local);
-
-				TrySetDataContextFromContent(Content);
-			}
-
-			return true;
-		}
-
-		return false;
-	}
-
-	void IFrameworkTemplatePoolAware.OnTemplateRecycled()
-	{
-		// This needs to be cleared on recycle, to prevent
-		// SetUpdateTemplate from being skipped in OnLoaded.
-		_firstLoadResetDone = false;
-	}
-
 	protected override void OnVisibilityChanged(Visibility oldValue, Visibility newValue)
 	{
 		base.OnVisibilityChanged(oldValue, newValue);
 
+#if !UNO_HAS_ENHANCED_LIFECYCLE
 		if (oldValue == Visibility.Collapsed && newValue == Visibility.Visible)
 		{
 			SetUpdateTemplate();
 		}
-	}
-
-	public void UpdateContentTemplateRoot()
-	{
-		if (Visibility == Visibility.Collapsed)
-		{
-			return;
-		}
-
-		//If ContentTemplateRoot is null, it must be updated even if the templates haven't changed
-		if (ContentTemplateRoot == null)
-		{
-			_dataTemplateUsedLastUpdate = null;
-		}
-
-		//ContentTemplate/ContentTemplateSelector will only be applied to a control with no Template, normally the innermost element
-		var dataTemplate = this.ResolveContentTemplate();
-
-		//Only apply template if it has changed
-		if (!object.Equals(dataTemplate, _dataTemplateUsedLastUpdate))
-		{
-			_dataTemplateUsedLastUpdate = dataTemplate;
-			ContentTemplateRoot = dataTemplate?.LoadContentCached() ?? Content as View;
-			if (ContentTemplateRoot != null)
-			{
-				IsUsingDefaultTemplate = false;
-			}
-		}
-
-		if (Content != null
-			&& !(Content is View)
-			&& ContentTemplateRoot == null
-		)
-		{
-			// Use basic default root for non-View Content if no template is supplied
-			SetContentTemplateRootToPlaceholder();
-		}
-
-		if (ContentTemplateRoot == null && Content is View contentView && dataTemplate == null)
-		{
-			// No template and Content is a View, set it directly as root
-			ContentTemplateRoot = contentView as View;
-		}
-
-		IsUsingDefaultTemplate = ContentTemplateRoot is ImplicitTextBlock;
+#endif
 	}
 
 	private void SetContentTemplateRootToPlaceholder()
@@ -1397,12 +1204,4 @@ public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePool
 #if !UNO_HAS_BORDER_VISUAL
 	private void UpdateBorder() => _borderRenderer.Update();
 #endif
-
-	private void SetUpdateTemplate()
-	{
-		UpdateContentTemplateRoot();
-		SetUpdateTemplatePartial();
-	}
-
-	partial void SetUpdateTemplatePartial();
 }
